% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_agents.R
\name{add_agents}
\alias{add_agents}
\title{Add AI agents}
\usage{
add_agents(x)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Add AI agents
}
\seealso{
Other both:
\code{\link[=add_checklist_badges]{add_checklist_badges()}},
\code{\link[=add_issue_templates]{add_issue_templates()}},
\code{\link[=check_filename]{check_filename()}},
\code{\link[=check_lintr]{check_lintr()}},
\code{\link[=check_spelling]{check_spelling()}},
\code{\link[=custom_dictionary]{custom_dictionary()}},
\code{\link[=print.checklist_spelling]{print.checklist_spelling()}},
\code{\link[=read_checklist]{read_checklist()}},
\code{\link[=update_citation]{update_citation()}},
\code{\link[=write_checklist]{write_checklist()}}
}
\concept{both}
