% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_checklist_badges.R
\name{add_checklist_badges}
\alias{add_checklist_badges}
\title{add checklist badges to a README}
\usage{
add_checklist_badges(x = ".", ...)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{...}{Additional arguments}
}
\description{
\itemize{
\item \code{check_package}: add a package check badge
\item \code{check_project}: add a project check badge
}
}
\examples{
\dontrun{
  add_badges(check_project = "inbo/checklist")
  add_badges(check_package = "inbo/checklist")
}
}
\seealso{
Other both:
\code{\link[=check_filename]{check_filename()}},
\code{\link[=check_lintr]{check_lintr()}},
\code{\link[=check_spelling]{check_spelling()}},
\code{\link[=custom_dictionary]{custom_dictionary()}},
\code{\link[=print.checklist_spelling]{print.checklist_spelling()}},
\code{\link[=read_checklist]{read_checklist()}},
\code{\link[=update_citation]{update_citation()}},
\code{\link[=write_checklist]{write_checklist()}}
}
\concept{both}
