% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_issue_templates.R
\name{add_issue_templates}
\alias{add_issue_templates}
\title{Add issue and pull request templates}
\usage{
add_issue_templates(x)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Add issue and pull request templates
}
\seealso{
Other both:
\code{\link[=add_agents]{add_agents()}},
\code{\link[=add_checklist_badges]{add_checklist_badges()}},
\code{\link[=check_filename]{check_filename()}},
\code{\link[=check_lintr]{check_lintr()}},
\code{\link[=check_spelling]{check_spelling()}},
\code{\link[=custom_dictionary]{custom_dictionary()}},
\code{\link[=print.checklist_spelling]{print.checklist_spelling()}},
\code{\link[=read_checklist]{read_checklist()}},
\code{\link[=update_citation]{update_citation()}},
\code{\link[=write_checklist]{write_checklist()}}
}
\concept{both}
