% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_maintainer.R
\name{ask_rightsholder_funder}
\alias{ask_rightsholder_funder}
\title{Ask for rights holder or funder}
\usage{
ask_rightsholder_funder(org, type = c("rightsholder", "funder"))
}
\arguments{
\item{org}{Organisation object}

\item{type}{Character, either \code{"rightsholder"} or \code{"funder"}}
}
\value{
A list with the selected names and the updated organisation object
}
\description{
Ask for rights holder or funder
}
\seealso{
Other utils: 
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{install_pak}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
