% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cran.R
\name{check_cran}
\alias{check_cran}
\title{Run all the package checks required by CRAN}
\usage{
check_cran(x = ".", quiet = FALSE, time_out = 15)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{quiet}{Whether to print check output during checking.}

\item{time_out}{The time in seconds to wait for a response from the world
clock API.
Default is 15 seconds, but you can increase this if you have a slow internet
connection.
If the world clock API is not available, the system clock will be used
without a warning, but the check will be less reliable.}
}
\value{
A \code{checklist} object.
}
\description{
CRAN imposes an impressive list of tests on every package before publication.
This suite of test is available in every R installation.
Hence we use this full suite of tests too.
Notice that \code{check_package()} runs several additional tests.
}
\seealso{
Other package:
\code{\link[=check_codemeta]{check_codemeta()}},
\code{\link[=check_description]{check_description()}},
\code{\link[=check_documentation]{check_documentation()}},
\code{\link[=check_environment]{check_environment()}},
\code{\link[=check_license]{check_license()}},
\code{\link[=check_package]{check_package()}},
\code{\link[=tidy_desc]{tidy_desc()}}
}
\concept{package}
