% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_description.R
\name{check_description}
\alias{check_description}
\title{Check the DESCRIPTION file}
\usage{
check_description(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the package.
Defaults to \code{.}.}
}
\description{
Check the DESCRIPTION file
}
