% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_documentation.R
\name{check_documentation}
\alias{check_documentation}
\title{Check the documentation}
\usage{
check_documentation(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the package.
Defaults to \code{.}.}
}
\description{
Check the documentation
}
