% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_environment.R
\name{check_environment}
\alias{check_environment}
\title{Make sure that the required environment variables are set on GitHub}
\usage{
check_environment(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
An invisible \code{Checklist} object.
}
\description{
Some actions will fail when these environment variables are not set.
This function does only work on GitHub.
}
