% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename.R
\name{check_filename}
\alias{check_filename}
\title{Check the style of file names}
\usage{
check_filename(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the package.
Defaults to \code{.}.}
}
\description{
Check the style of file names
}
