% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_folder.R
\name{check_folder}
\alias{check_folder}
\title{Check the folder structure}
\usage{
check_folder(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
For the time being, this function only checks projects.
Keep in mind that R packages have requirements for the folder structure.
That is one of things that \code{check_cran()} checks.
}
\section{Recommended folder structure}{
\itemize{
\item \code{source}: contains all \code{R} scripts and \code{Rmd} files.
\item \code{data}: contains all data files.
}
}

\section{\code{source}}{
A simple project with only \code{R} scripts or only \code{Rmd} files can place all the
files directly in the \code{source} folder.

More elaborate projects should place in the files in several folders under
\code{source}.
Place every \code{bookdown} document in a dedicated folder.
And create an RStudio project for that folder.
}

\section{\code{data}}{
Simple projects in which \code{source} has no subfolders, place \code{data} at the
root of the project.
For more elaborate project you must choose between either \code{data} at the root
of the project or \code{data} as subfolder of the subfolders of \code{source}.
E.g. \code{source/report/data}.

Place the data in an open file format.
E.g. \code{csv}, \code{txt} or \code{tsv} for tabular data.
We strongly recommend to use \code{git2rdata::write_vc()} to store such data.
Use the \href{https://www.geopackage.org/}{\code{geopackage}} format for spatial data.
Optionally add description of the data as markdown files.
}

\seealso{
Other project: 
\code{\link{check_project}()},
\code{\link{check_source}()}
}
\concept{project}
