% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_description.R
\name{check_license}
\alias{check_license}
\title{Check the license}
\usage{
check_license(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Check the license
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()}
}
\concept{package}
