% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lintr.R
\name{check_lintr}
\alias{check_lintr}
\title{Check the packages for linters}
\usage{
check_lintr(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the package.
Defaults to \code{.}.}
}
\description{
Check the packages for linters
}
