% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_project.R
\name{check_project}
\alias{check_project}
\title{Run the required quality checks on a project}
\usage{
check_project(x = ".", fail = !interactive(), quiet = FALSE)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{fail}{Should the function return an error in case of a problem?
Defaults to \code{TRUE} on a non-interactive session and \code{FALSE} on an interactive
session.}

\item{quiet}{Whether to print check output during checking.}
}
\description{
Set or update the required checks via \code{setup_project()}.
}
\seealso{
Other project: 
\code{\link{check_folder}()},
\code{\link{check_source}()}
}
\concept{project}
