% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rmd.R
\name{check_rmd_renders}
\alias{check_rmd_renders}
\title{Check an Rmd file renders using rmarkdown}
\usage{
check_rmd_renders(
  file,
  install_missing = FALSE,
  update_packages = FALSE,
  output_dir = NULL,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{file}{Path of an Rmd file}

\item{install_missing}{Should any missing packages be installed. Default \code{FALSE}.}

\item{update_packages}{Should installed packages be updated, using
\code{\link[=update_installed_pkgs]{update_installed_pkgs()}}, before rendering. Default \code{FALSE}.}

\item{output_dir}{Directory for the rendered output, passed to
\code{\link[rmarkdown:render]{rmarkdown::render()}}. The default (\code{NULL}) renders next to the input file.}

\item{quiet}{Should knitting and pandoc output be suppressed. Default \code{TRUE}.}

\item{...}{Additional arguments to pass to \code{render()}}
}
\value{
The path of the rendered output file, as returned by
\code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Check an Rmd file renders using rmarkdown
}
\examples{
\dontrun{
check_rmd_renders("hw1.Rmd")
}

}
