% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_source.R
\name{check_source}
\alias{check_source}
\title{Standardised test for an R source repository}
\usage{
check_source(x = ".", fail = !interactive())
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{fail}{Should the function return an error in case of a problem?
Defaults to \code{TRUE} on non-interactive session and \code{FALSE} on an interactive
session.}
}
\description{
Defunct function.
Please use \code{check_project()} instead.
}
\seealso{
Other project: 
\code{\link{check_folder}()},
\code{\link{check_project}()}
}
\concept{project}
