% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spelling.R
\name{check_spelling}
\alias{check_spelling}
\title{Spell check a package or project}
\usage{
check_spelling(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Spell check a package or project
}
