% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_template.R
\name{checklist_template}
\alias{checklist_template}
\title{Create a template with acceptable problems in the package}
\usage{
checklist_template(output, path = ".")
}
\arguments{
\item{output}{The output of \code{check_package()}.}

\item{path}{The path of the package.
Defaults to \code{"."}}
}
\description{
Create a template with acceptable problems in the package
}
