% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_meta_class.R
\name{citation_meta}
\alias{citation_meta}
\title{The \code{citation_meta} R6 class}
\description{
A class which contains citation information.
}
\seealso{
Other class: 
\code{\link{checklist}},
\code{\link{org_item}},
\code{\link{org_list}},
\code{\link{organisation}},
\code{\link{spelling}}
}
\concept{class}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_errors}}{Return the errors}

\item{\code{get_meta}}{Return the meta data as a list}

\item{\code{get_notes}}{Return the notes}

\item{\code{get_person}}{Return the authors and organisations as a list of
\code{person} objects.}

\item{\code{get_type}}{A string indicating the type of source.}

\item{\code{get_path}}{The path to the project.}

\item{\code{get_warnings}}{Return the warnings}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-citation_meta-new}{\code{citation_meta$new()}}
\item \href{#method-citation_meta-print}{\code{citation_meta$print()}}
\item \href{#method-citation_meta-clone}{\code{citation_meta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-citation_meta-new"></a>}}
\if{latex}{\out{\hypertarget{method-citation_meta-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{citation_meta} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{citation_meta$new(path = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to the root of the project.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-citation_meta-print"></a>}}
\if{latex}{\out{\hypertarget{method-citation_meta-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{citation_meta} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{citation_meta$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{currently ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-citation_meta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-citation_meta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{citation_meta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
