% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_external_functions.R
\name{cl_blend}
\alias{cl_blend}
\title{Blend Checklist Colors with RStudio Theme}
\usage{
cl_blend(cl)
}
\arguments{
\item{cl}{A checklist object}
}
\description{
This function is provided to help blend a checklist in
with the RStudio environment in cases where the color scheme was
changed or a new theme was selected in RStudio.
}
\examples{
# start with a basic checklist
my_checklist <- cl_create()

# change background color to orange
cl_set(my_checklist, background.color = "orange")

# blend it back to RStudio theme
cl_blend(my_checklist)

}
