% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_external_functions.R
\name{cl_remind}
\alias{cl_remind}
\title{Create a Reminder File for a Checklist}
\usage{
cl_remind(cl)
}
\arguments{
\item{cl}{A checklist object}
}
\description{
This function creates a .txt file that will attempt to
show up in your project/working directory, as well as the
editor in RStudio. The primary inspiration for this function was
to provide an added file that would show up in a git diff to
prevent commits prior to completing all expected work.
}
\examples{
# remind at the time of creation
my_reminding_checklist <- cl_create(remind = TRUE)

# or make the file afterwards
my_basic_checklist <- cl_create()
cl_remind(my_basic_checklist)

}
