% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_external_functions.R
\name{cl_save}
\alias{cl_save}
\title{Save a Checklist Object}
\usage{
cl_save(cl, file)
}
\arguments{
\item{cl}{A Checklist object to save}

\item{file}{A String containing the file path of the checklist.
Needs to be an .rds file.}
}
\description{
This function allows you to save your checklist to a more
permanent place than R's temp directory. It will check to make sure
the supplied filename ends with .rds (or append .rds if no extension)
is detected, but is oterwise just a wrapper for saveRDS.
}
\examples{
# create a checklist
my_checklist <- cl_create()

# save it
cl_save(my_checklist, "my_checklist.rds")

# remove it from workspace
rm(cl)

# load it back in
cl <- cl_load("my_checklist.rds")

}
