% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_git.R
\name{clean_git}
\alias{clean_git}
\title{Clean the git repository}
\usage{
clean_git(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{A path to an existing local git repository.}

\item{verbose}{Print information each time a reference is updated
locally. Default is \code{TRUE}.}
}
\description{
\itemize{
\item update local branches that are behind their counterpart on origin.
\item list local branches that have diverged from their counterpart the origin.
\item list local branches without counterpart on origin that have diverged from
the main branch.
\item remove local branches without counterpart on origin and fully merged into
the main branch.
\item remove local copies of origin branches deleted at the origin.
}
}
\seealso{
Other utils: 
\code{\link{create_hexsticker}()},
\code{\link{is_workdir_clean}()},
\code{\link{orcid2person}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
