% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_draft_pr.R
\name{create_draft_pr}
\alias{create_draft_pr}
\title{Create a draft pull request}
\usage{
create_draft_pr(x = ".")
}
\arguments{
\item{x}{The path to the package.}
}
\description{
This function creates a draft pull request for the current version of the
package.
}
\seealso{
Other git: 
\code{\link{clean_git}()},
\code{\link{get_branches_tags}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{new_branch}()},
\code{\link{set_tag}()}
}
\concept{git}
