% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexsticker.R
\name{create_hexsticker}
\alias{create_hexsticker}
\title{Make hexagonal logo for package}
\usage{
create_hexsticker(package_name, filename = "man/figures/hexsticker.png")
}
\arguments{
\item{package_name}{package name that should be mentioned on the hexsticker}

\item{filename}{filename to save the sticker (default saved in man/figures)}
}
\value{
A figure is saved in the working directory or provided path
}
\description{
This function makes a hexsticker in INBO style for the provided
package name.
It requires the package hexSticker to be installed.
}
\examples{
# make tempfile to save logo (or just use (path and) filename)
output <- tempfile(pattern = "hexsticker", fileext = ".png")
create_hexsticker("checklist", filename = output)
}
\seealso{
Other utils: 
\code{\link{is_workdir_clean}()},
\code{\link{orcid2person}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
