% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create an R package according to INBO requirements}
\usage{
create_package(package, path = ".")
}
\arguments{
\item{package}{Name of the new package.}

\item{path}{Where to create the package directory.}
}
\description{
Creates a package template in a new folder.
Use this function when you want to start a new package.
Please DO READ \code{vignette("getting_started")} before running this function.
}
\details{
What you get with a checklist setup:
\itemize{
\item minimal folder structure and files required for an R package using INBO
guidelines with GPL-3 or MIT license.
\item an RStudio project file
\item a local git repository
\item an initial \code{NEWS.md} file
\item a template for an \code{README.Rmd}
\item set-up for automated checks and releases of the package using GitHub
Actions.
\item a code of conduct and contributing guidelines.
\item the set-up for a \code{pkgdown} website using the INBO corporate identity.
}
}
\seealso{
Other setup: 
\code{\link{create_project}()},
\code{\link{new_org_list}()},
\code{\link{prepare_ghpages}()},
\code{\link{set_license}()},
\code{\link{setup_package}()},
\code{\link{setup_project}()},
\code{\link{setup_source}()}
}
\concept{setup}
