% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{create_project}
\alias{create_project}
\title{Initialise a new R project}
\usage{
create_project(path, project)
}
\arguments{
\item{path}{The folder in which to create the project as a folder.}

\item{project}{The name of the project.}
}
\description{
This function creates a new RStudio project with \code{checklist} functionality.
}
\seealso{
Other setup: 
\code{\link{create_package}()},
\code{\link{prepare_ghpages}()},
\code{\link{setup_package}()},
\code{\link{setup_project}()},
\code{\link{setup_source}()}
}
\concept{setup}
