% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_dictionary.R
\name{custom_dictionary}
\alias{custom_dictionary}
\title{Add words to custom dictionaries}
\usage{
custom_dictionary(issues)
}
\arguments{
\item{issues}{The output of \code{check_spelling()}.}
}
\description{
Add words to custom dictionaries
}
\seealso{
Other both:
\code{\link[=add_agents]{add_agents()}},
\code{\link[=add_checklist_badges]{add_checklist_badges()}},
\code{\link[=check_filename]{check_filename()}},
\code{\link[=check_lintr]{check_lintr()}},
\code{\link[=check_spelling]{check_spelling()}},
\code{\link[=print.checklist_spelling]{print.checklist_spelling()}},
\code{\link[=read_checklist]{read_checklist()}},
\code{\link[=update_citation]{update_citation()}},
\code{\link[=write_checklist]{write_checklist()}}
}
\concept{both}
