% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_organisation.R
\name{default_organisation}
\alias{default_organisation}
\title{Write default organisation settings}
\usage{
default_organisation(org = organisation$new())
}
\arguments{
\item{org}{An \code{organisation} object.
Create it with \code{organisation$new()}.}
}
\description{
Store the default organisation rules.
First run \code{org <- organisation$new()} with the appropriate argument.
Next you can store the configuration with \code{default_organisation(org)}.
}
\seealso{
Other both: 
\code{\link{add_badges}()},
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{custom_dictionary}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{update_citation}()},
\code{\link{write_checklist}()},
\code{\link{write_citation_cff}()},
\code{\link{write_organisation}()},
\code{\link{write_zenodo_json}()}
}
\concept{both}
