% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execshell}
\alias{execshell}
\title{Pass command lines to a shell}
\usage{
execshell(commandstring, intern = FALSE, path = ".", ...)
}
\arguments{
\item{commandstring}{The system command to be invoked, as a string.
Multiple commands can be combined in this single string, e.g. with a
multiline string.}

\item{intern}{a logical (not \code{NA}) which indicates whether to
    capture the output of the command as an \R character vector.}

\item{path}{The path from where the command string needs to be executed}

\item{...}{Other arguments passed to \code{\link[base]{system}} or
\code{\link[base]{shell}}.}
}
\description{
Cross-platform function to pass a command to the shell, using either
\code{\link[base]{system}} or
(Windows-only) \code{\link[base]{shell}}, depending on the operating system.
}
\keyword{internal}
