% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches_tags.R
\name{get_branches_tags}
\alias{get_branches_tags}
\title{Get branches and tags of a GitHub repository}
\usage{
get_branches_tags(owner, repo)
}
\arguments{
\item{owner}{Repository owner}

\item{repo}{Repository name}
}
\value{
A sorted character vector of branch and tag names, excluding the
\code{gh-pages} branch.
}
\description{
Get branches and tags of a GitHub repository
}
\seealso{
Other git: 
\code{\link{clean_git}()},
\code{\link{create_draft_pr}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{new_branch}()},
\code{\link{set_tag}()}
}
\concept{git}
