% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_org_list.R
\name{get_default_org_list}
\alias{get_default_org_list}
\title{Get the default organization list}
\usage{
get_default_org_list(x = ".")
}
\arguments{
\item{x}{The path to the repository.
Defaults to the current working directory.}
}
\value{
An \code{org_list} object containing the organisation list.
The function also stores the information in the user's R configuration.
}
\description{
This function retrieves the default organisation list from the
\code{organisation.yml} file in the organisations \code{checklist} repository.
The \code{origin} of the repository is used to determine the root URL of the
organisation.
}
\seealso{
Other utils: 
\code{\link{ask_rightsholder_funder}()},
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{inbo_org_list}()},
\code{\link{install_pak}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
