% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit_on_failure.R
\name{handle_error}
\alias{handle_error}
\title{Handle runtime errors}
\usage{
handle_error(
  expr,
  on_success = {
 },
  on_error = {
 },
  on_warning = {
 },
  finally = {
 }
)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{on_success}{Expression to evaluate if \code{expr} succeeds}

\item{on_error}{Expression to evaluate if \code{expr} generates an error}

\item{on_warning}{Expression to evaluate if \code{expr} generates a warning}

\item{finally}{Expression to evaluate after \code{expr} and \code{on_success} or \code{on_error} or \code{on_warning},
usually used for cleanup.}
}
\value{
Invisible result of \code{expr}
}
\description{
Evaluates the given expression and then runs the \code{on_success}, \code{on_error}, or \code{on_warning}
expressions depending on the result of the initial evaluation.
}
