% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_list_class.R
\name{inbo_org_list}
\alias{inbo_org_list}
\title{The INBO organisation list}
\usage{
inbo_org_list()
}
\description{
The INBO organisation list
}
\seealso{
Other utils: 
\code{\link{ask_rightsholder_funder}()},
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{get_default_org_list}()},
\code{\link{install_pak}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
