% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_repository}
\alias{is_repository}
\title{Determine if a directory is in a git repository}
\usage{
is_repository(path = ".")
}
\arguments{
\item{path}{the location of the git repository, see details.}
}
\value{
TRUE if directory is in a git repository else FALSE
}
\description{
The path arguments specifies the directory at which to start the search for
a git repository.
If it is not a git repository itself, then its parent directory is consulted,
then the parent's parent, and so on.
}
\seealso{
Other git: 
\code{\link{clean_git}()},
\code{\link{create_draft_pr}()},
\code{\link{is_workdir_clean}()},
\code{\link{new_branch}()},
\code{\link{set_tag}()}
}
\concept{git}
