% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tracked_not_modified}
\alias{is_tracked_not_modified}
\title{Check if a file is tracked and not modified}
\usage{
is_tracked_not_modified(file, repo = ".")
}
\arguments{
\item{file}{path relative to the git root directory.}

\item{repo}{path to the repository}
}
\description{
Check if a file is tracked and not modified
}
\keyword{internal}
