% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_workdir_clean}
\alias{is_workdir_clean}
\title{Check if the current workdir of a repo is clean}
\usage{
is_workdir_clean(repo)
}
\arguments{
\item{repo}{Either a \code{git2r::repository()} or path to the repository.}
}
\value{
\code{TRUE} when there are no staged, unstaged or untracked files.
Otherwise \code{FALSE}
}
\description{
A clean working directory has no staged, unstaged or untracked files.
}
\seealso{
Other utils: 
\code{\link{orcid2person}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
