% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_external_functions.R
\name{item-functions}
\alias{item-functions}
\alias{cl_add}
\alias{cl_edit}
\alias{cl_complete}
\alias{cl_uncomplete}
\alias{cl_remove}
\title{Checklist Item Functions}
\usage{
cl_add(cl, text)

cl_edit(cl, at_items, text)

cl_complete(cl, at_items)

cl_uncomplete(cl, at_items)

cl_remove(cl, at_items)
}
\arguments{
\item{cl}{A checklist object}

\item{text}{A String or character vector containing text to add
or edit}

\item{at_items}{An Integer or Integer vector dectermining where
to edit, complete, uncomplete, or remove items}
}
\description{
This series of functions allows you to manipulate items
in a checklist more intuitively and conveniently than using cl_set.
They are designed to focus on manipulating a single task at a time.
If you need to manipultate several items at once, consider using
cl_set with a character vector as the 'tasks' argument and a
logical vector as the 'completed' argument
}
\examples{
# create a checklist
cl <- cl_create()

# add a few items
cl_add(cl, "item 1")
cl_add(cl, "item 2")
cl_add(cl, "item 3")

# edit the second item
cl_edit(cl, 2, "update item 2")

# complete the second item
cl_complete(cl, 2)

# uncomplete the second item
cl_uncomplete(cl, 2)

# remove the second item
cl_remove(cl, 2)

}
