% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_author.R
\name{menu_first}
\alias{menu_first}
\title{Improved version of menu()}
\usage{
menu_first(choices, graphics = FALSE, title = NULL)
}
\arguments{
\item{choices}{a character vector of choices}

\item{graphics}{a logical indicating whether a graphics menu should be
    used if available.}

\item{title}{a character string to be used as the title of the menu.
    \code{NULL} is also accepted.}
}
\description{
Improved version of menu()
}
\seealso{
Other utils: 
\code{\link{ask_rightsholder_funder}()},
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{install_pak}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
