% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_branch.R
\name{new_branch}
\alias{new_branch}
\title{Create a new branch after cleaning the repo}
\usage{
new_branch(branch, path = ".", verbose = TRUE, force = FALSE)
}
\arguments{
\item{branch}{Name of the new branch}

\item{path}{A path to an existing local git repository.}

\item{verbose}{Print information each time a reference is updated
locally. Default is \code{TRUE}.}

\item{force}{Overwrite existing branch. Default = FALSE}
}
\description{
First run \code{clean_git()}.
Then create the new branch from the (updated) main branch.
}
\seealso{
Other utils: 
\code{\link{clean_git}()},
\code{\link{create_hexsticker}()},
\code{\link{is_workdir_clean}()},
\code{\link{orcid2person}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
