% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_org_list.R
\name{new_org_list}
\alias{new_org_list}
\title{Interactively create a new organisation list}
\usage{
new_org_list()
}
\description{
An interactive alternative for \code{org_list$new()}.
Reuses available organisations where possible.
}
\seealso{
\code{\link{org_list}}, \code{\link{org_item}}
}
