% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orcid2person}
\alias{orcid2person}
\title{Convert an ORCID to a \code{person} object.}
\usage{
orcid2person(orcid, email, role = c("aut", "cre"))
}
\arguments{
\item{orcid}{The ORCID of the person.}

\item{email}{An optional email of the person.
Require when the ORCID record does not contain a public email.}

\item{role}{The role of the person.
See \code{utils::person} for all possible values.}
}
\description{
Convert an ORCID to a \code{person} object.
}
