% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_item_class.R
\name{org_item}
\alias{org_item}
\title{The \code{org_item}  R6 class}
\description{
A class containing a single organisation
}
\seealso{
Other class: 
\code{\link{checklist}},
\code{\link{citation_meta}},
\code{\link{org_list}},
\code{\link{organisation}},
\code{\link{spelling}}
}
\concept{class}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{as_list}}{The organisation as a list.}

\item{\code{get_community}}{The organisation Zenodo community.}

\item{\code{get_default_name}}{The organisation default name.}

\item{\code{get_email}}{The organisation email.}

\item{\code{get_funder}}{The funder rules.}

\item{\code{get_git}}{The organisational root git URL.}

\item{\code{get_name}}{The organisation names.}

\item{\code{get_orcid}}{The ORCID rules.}

\item{\code{get_rightsholder}}{The rightsholder rules.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-org_item-new}{\code{org_item$new()}}
\item \href{#method-org_item-as_person}{\code{org_item$as_person()}}
\item \href{#method-org_item-print}{\code{org_item$print()}}
\item \href{#method-org_item-clone}{\code{org_item$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_item-new"></a>}}
\if{latex}{\out{\hypertarget{method-org_item-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{org_item} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_item$new(
  name,
  email,
  orcid = FALSE,
  rightsholder = c("optional", "single", "shared", "when no other"),
  funder = c("optional", "single", "shared", "when no other"),
  git,
  ror,
  zenodo
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A named vector with the organisation names in one or more
languages.
The first item in the vector is the default language.
The names of the vector must match the language code.}

\item{\code{email}}{An optional email address for the organisation.
Used to contact the organisation.}

\item{\code{orcid}}{Whether the organisation requires an ORCID for every person
that uses this organisation as affiliation.}

\item{\code{rightsholder}}{The required copyright holder status for the
organisation.
\code{"optional"} means that the organisation is not required as the copyright
holder.
\code{"single"} means that the organisation must be the only copyright holder.
\code{"shared"} means that the organisation must be one of the copyright
holders.
\code{"no other"} means that if no other copyright holder is specified,
the organisation must be the copyright holder.}

\item{\code{funder}}{The required funder status for the organisation.
The categories are the same as for \code{rightsholder}.}

\item{\code{git}}{The optional root URL to the organisations git repositories.
The organisation with matching URL in the git remote will be used as
the primary organisation.}

\item{\code{ror}}{The optional ROR ID of the organisation.}

\item{\code{zenodo}}{The optional Zenodo community ID of the organisation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_item-as_person"></a>}}
\if{latex}{\out{\hypertarget{method-org_item-as_person}{}}}
\subsection{Method \code{as_person()}}{
as_person The organisation as a person.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_item$as_person(lang = names(private$name)[1], role = c("cph", "fnd"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{The language to use for the organisation name.
Defaults to the first language in the \code{name} vector.}

\item{\code{role}}{The role of the person in the organisation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_item-print"></a>}}
\if{latex}{\out{\hypertarget{method-org_item-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{org_item} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_item$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_item-clone"></a>}}
\if{latex}{\out{\hypertarget{method-org_item-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_item$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
