% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_list_class.R
\name{org_list}
\alias{org_list}
\title{The \code{org_list}  R6 class}
\description{
A class containing a list of organisations
}
\seealso{
Other class: 
\code{\link{checklist}},
\code{\link{citation_meta}},
\code{\link{org_item}},
\code{\link{organisation}},
\code{\link{spelling}}
}
\concept{class}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_default_name}}{The organisations default name.}

\item{\code{get_email}}{The organisations email.}

\item{\code{get_languages}}{The different languages of the organisations.}

\item{\code{which_funder}}{The required rightsholders.}

\item{\code{which_rightsholder}}{The required rightsholders.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-org_list-get_person}{\code{org_list$get_person()}}
\item \href{#method-org_list-get_zenodo_by_email}{\code{org_list$get_zenodo_by_email()}}
\item \href{#method-org_list-get_match}{\code{org_list$get_match()}}
\item \href{#method-org_list-get_name_by_domain}{\code{org_list$get_name_by_domain()}}
\item \href{#method-org_list-new}{\code{org_list$new()}}
\item \href{#method-org_list-print}{\code{org_list$print()}}
\item \href{#method-org_list-read}{\code{org_list$read()}}
\item \href{#method-org_list-validate_person}{\code{org_list$validate_person()}}
\item \href{#method-org_list-validate_rules}{\code{org_list$validate_rules()}}
\item \href{#method-org_list-write}{\code{org_list$write()}}
\item \href{#method-org_list-clone}{\code{org_list$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-get_person"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-get_person}{}}}
\subsection{Method \code{get_person()}}{
Return the organisation with matching email as a \code{person()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$get_person(email, role = c("cph", "fnd"), lang)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email address of the organisation.}

\item{\code{role}}{The role of the person to return.}

\item{\code{lang}}{The language to return the organisation name in.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-get_zenodo_by_email"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-get_zenodo_by_email}{}}}
\subsection{Method \code{get_zenodo_by_email()}}{
Return a vector of Zenodo communities associated with the
organisations with matching email.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$get_zenodo_by_email(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email addresses to match against.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector with the communities.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-get_match"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-get_match}{}}}
\subsection{Method \code{get_match()}}{
Return the organisation with a matching name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$get_match(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the organisation to match.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the organisation name, email and match ratio.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-get_name_by_domain"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-get_name_by_domain}{}}}
\subsection{Method \code{get_name_by_domain()}}{
Return the organisation names with a matching email domain.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$get_name_by_domain(email, lang)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email address to match the domain against.}

\item{\code{lang}}{The language to return the organisation name in.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function extracts the domain from the email address and
matches it against the organisation email addresses.
If multiple organisations have the same domain,
the function returns all matching names.
If the language is not available for a specific organisation,
it will return the first available name.
}

\subsection{Returns}{
A character vector with the organisation names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-new"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{org_list} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more \code{org_item} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-print"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{org_list} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-read"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-read}{}}}
\subsection{Method \code{read()}}{
Read the \code{org_list} object from an \code{organisation.yml} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$read(x = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The path to the directory where the \code{organisation.yml} file
is stored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-validate_person"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-validate_person}{}}}
\subsection{Method \code{validate_person()}}{
Validate a \code{person} object given the \code{org_list} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$validate_person(person, lang)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{person}}{The \code{person} object to validate.}

\item{\code{lang}}{The language to use for affiliation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-validate_rules"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-validate_rules}{}}}
\subsection{Method \code{validate_rules()}}{
Validate the rules for the rightsholder and funder.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$validate_rules(rightsholder = person(), funder = person())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rightsholder}}{The rightsholders as a \code{person} object.}

\item{\code{funder}}{The funders as a \code{person} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-write"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-write}{}}}
\subsection{Method \code{write()}}{
Write the \code{org_list} object to an \code{organisation.yml} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$write(x = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The path to the directory where the \code{organisation.yml} file
should be written.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-clone"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
