% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ghpages.R
\name{prepare_ghpages}
\alias{prepare_ghpages}
\title{Prepare a \code{gh-pages} branch with a place holder page}
\usage{
prepare_ghpages(x = ".", verbose = TRUE)
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{verbose}{display some progress info while downloading}
}
\description{
Prepare a \code{gh-pages} branch with a place holder page
}
\seealso{
Other setup: 
\code{\link{create_package}()},
\code{\link{setup_package}()},
\code{\link{setup_source}()}
}
\concept{setup}
