% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spelling.R
\name{print.checklist_spelling}
\alias{print.checklist_spelling}
\title{Display a \code{checklist_spelling} summary}
\usage{
\method{print}{checklist_spelling}(x, ...)
}
\arguments{
\item{x}{The \code{checklist_spelling} object}

\item{...}{currently ignored}
}
\description{
Display a \code{checklist_spelling} summary
}
\seealso{
Other both:
\code{\link[=add_checklist_badges]{add_checklist_badges()}},
\code{\link[=check_filename]{check_filename()}},
\code{\link[=check_lintr]{check_lintr()}},
\code{\link[=check_spelling]{check_spelling()}},
\code{\link[=custom_dictionary]{custom_dictionary()}},
\code{\link[=read_checklist]{read_checklist()}},
\code{\link[=update_citation]{update_citation()}},
\code{\link[=write_checklist]{write_checklist()}}
}
\concept{both}
