% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit_on_failure.R
\name{quit_on_failure}
\alias{quit_on_failure}
\title{Quit on Failure}
\usage{
quit_on_failure(expr, n_br = 1)
}
\arguments{
\item{expr}{Expression returning a logical vector}

\item{n_br}{Number of leading and trailing blank lines to print}
}
\value{
The value of \code{expr}, invisibly. If any element of the result is
not \code{TRUE} the R session is terminated with exit status 1.
}
\description{
Evaluate the given expression and if any of the returned values are \code{FALSE} then
exit with status 1.
}
\details{
This is needed when using \verb{check_*} functions within a continuous integration pipeline
to signal that a step failed.
}
\examples{
\dontrun{
quit_on_failure(check_required_files("hw1.Rmd"))
}

}
