% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_organisation.R
\name{read_organisation}
\alias{read_organisation}
\title{Read the organisation file}
\usage{
read_organisation(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
An \code{organisation} object.
}
\description{
The checklist package stores organisation information in the
\code{organisation.yml} file in the root of a project.
}
\seealso{
Other both: 
\code{\link{add_badges}()},
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{custom_dictionary}()},
\code{\link{default_organisation}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{update_citation}()},
\code{\link{write_checklist}()},
\code{\link{write_citation_cff}()},
\code{\link{write_organisation}()},
\code{\link{write_zenodo_json}()}
}
\concept{both}
