% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-rfe.R
\name{rfe}
\alias{rfe}
\title{Recursive Feature Elimination}
\usage{
rfe(x, y, get_loss, prop = 1, k = 1, verbose = T)
}
\arguments{
\item{x}{The independent variables (called features). Typically a matrix.}

\item{y}{The dependent variables. Typically a vector.}

\item{get_loss}{A loss function that takes the arguments 'x' and 'y' and
returns a numeric value.}

\item{prop}{The proportion of samples used to compute loss. Provide a
numeric value between 0 and 1.}

\item{k}{k The number of features eliminated each step. Provide an integer
greater than 0.}

\item{verbose}{A boolean. Toggles whether to print progress.}
}
\value{
A data.frame of ranked features.
}
\description{
Perform recursive feature elimination.
}
\details{
Feature elimination can be sped up by computing the loss on a
proportion ('prop') of the data, or by eliminating several ('k')
features per step.
}
