% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_license.R
\name{set_license}
\alias{set_license}
\title{Set the proper license}
\usage{
set_license(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Set the proper license
}
\seealso{
Other setup: 
\code{\link{create_package}()},
\code{\link{create_project}()},
\code{\link{prepare_ghpages}()},
\code{\link{setup_package}()},
\code{\link{setup_project}()},
\code{\link{setup_source}()}
}
\concept{setup}
