% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tag.R
\name{set_tag}
\alias{set_tag}
\title{Set a New Tag}
\usage{
set_tag(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
This function is a part of the GitHub Action.
Therefore it only works when run in a GitHub Action on the main or master
branch.
Otherwise it will only return a message.
It sets a new tag at the current commit using the related entry from
\code{NEWS.md} as message.
This tag will turn into a release.
}
\seealso{
Other git: 
\code{\link{clean_git}()},
\code{\link{create_draft_pr}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{new_branch}()}
}
\concept{git}
