% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_package.R
\name{setup_package}
\alias{setup_package}
\title{Add or update the checklist infrastructure to an existing package}
\usage{
setup_package(path = ".")
}
\arguments{
\item{path}{The path to the package.
Defaults to \code{"."}.}
}
\description{
Add or update the checklist infrastructure to an existing package
}
\seealso{
Other setup: 
\code{\link{create_package}()}
}
\concept{setup}
