% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Set-up \code{checklist} on an existing R project}
\usage{
setup_project(path = ".")
}
\arguments{
\item{path}{the project root folder}
}
\description{
Use this function to set-up or change the \code{checklist} infrastructure for an
existing project.
The function interactively asks questions to set-up the required checks.
}
\seealso{
Other setup: 
\code{\link{create_package}()},
\code{\link{create_project}()},
\code{\link{new_org_list}()},
\code{\link{prepare_ghpages}()},
\code{\link{set_license}()},
\code{\link{setup_package}()},
\code{\link{setup_source}()}
}
\concept{setup}
