% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_source.R
\name{setup_source}
\alias{setup_source}
\title{Add or update the checklist infrastructure to a repository with source files.}
\usage{
setup_source(path = ".")
}
\arguments{
\item{path}{The path to the project.
Defaults to \code{"."}.}
}
\description{
Add or update the checklist infrastructure to a repository with source files.
}
\seealso{
Other setup: 
\code{\link{create_package}()},
\code{\link{setup_package}()}
}
\concept{setup}
