% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling_class.R
\name{spelling}
\alias{spelling}
\title{The spelling R6 class}
\description{
A class with the configuration for spell checking
}
\seealso{
Other class: 
\code{\link{checklist}},
\code{\link{citation_meta}},
\code{\link{organisation}}
}
\concept{class}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{default}}{The default language of the project.}

\item{\code{get_md}}{The markdown files within the project.}

\item{\code{get_r}}{The R files within the project.}

\item{\code{get_rd}}{The Rd files within the project.}

\item{\code{settings}}{A list with current spell checking settings.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-spelling-new}{\code{spelling$new()}}
\item \href{#method-spelling-print}{\code{spelling$print()}}
\item \href{#method-spelling-set_default}{\code{spelling$set_default()}}
\item \href{#method-spelling-set_exceptions}{\code{spelling$set_exceptions()}}
\item \href{#method-spelling-set_ignore}{\code{spelling$set_ignore()}}
\item \href{#method-spelling-set_other}{\code{spelling$set_other()}}
\item \href{#method-spelling-clone}{\code{spelling$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spelling-new"></a>}}
\if{latex}{\out{\hypertarget{method-spelling-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{spelling} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spelling$new(language, base_path = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{language}}{the default language.}

\item{\code{base_path}}{the base path of the project}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spelling-print"></a>}}
\if{latex}{\out{\hypertarget{method-spelling-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{spelling} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spelling$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{currently ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spelling-set_default"></a>}}
\if{latex}{\out{\hypertarget{method-spelling-set_default}{}}}
\subsection{Method \code{set_default()}}{
Define which files to ignore or to spell check in a
different language.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spelling$set_default(language)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{language}}{The language.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spelling-set_exceptions"></a>}}
\if{latex}{\out{\hypertarget{method-spelling-set_exceptions}{}}}
\subsection{Method \code{set_exceptions()}}{
Define which files to ignore or to spell check in a
different language.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spelling$set_exceptions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spelling-set_ignore"></a>}}
\if{latex}{\out{\hypertarget{method-spelling-set_ignore}{}}}
\subsection{Method \code{set_ignore()}}{
Manually set the ignore vector.
Only use this if you known what you are doing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spelling$set_ignore(ignore)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{The character vector with ignore file patterns.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spelling-set_other"></a>}}
\if{latex}{\out{\hypertarget{method-spelling-set_other}{}}}
\subsection{Method \code{set_other()}}{
Manually set the other list.
Only use this if you known what you are doing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spelling$set_other(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{a list with file patterns per additional language.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spelling-clone"></a>}}
\if{latex}{\out{\hypertarget{method-spelling-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spelling$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
