% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_authors.R
\name{store_authors}
\alias{store_authors}
\title{Store author details for later usage}
\usage{
store_authors(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Store author details for later usage
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{install_pak}()},
\code{\link{menu_first}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
