% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_citation.R
\name{update_citation}
\alias{update_citation}
\title{Create or update the citation files}
\usage{
update_citation(x = ".", quiet = FALSE)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{quiet}{Whether to print check output during checking.}
}
\value{
An invisible \code{checklist} object.
}
\description{
The function extracts citation meta data from the project.
Then it checks the required meta data.
Upon success, it writes several files.
}
\details{
\itemize{
\item \code{.zenodo.json} contains the citation information in the format that
\href{https://zenodo.org}{Zenodo} requires.
\item \code{CITATION.cff} provides the citation information in the format that
\href{https://github.com}{GitHub} requires.
\item \code{inst/CITATION} provides the citation information in the format that
R packages require.
It is only relevant for packages.
}
}
\note{
Source of the citation meta data:
\itemize{
\item package: \code{DESCRIPTION}
\item project: \code{README.md}
}

Should you want to add more information to the \code{inst/CITATION} file,
add it to that file outside \verb{# begin checklist entry} and
\verb{# end checklist entry}.
}
\seealso{
Other both: 
\code{\link{add_badges}()},
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{custom_dictionary}()},
\code{\link{default_organisation}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{write_checklist}()},
\code{\link{write_citation_cff}()},
\code{\link{write_organisation}()},
\code{\link{write_zenodo_json}()}
}
\concept{both}
