% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_author.R
\name{use_author}
\alias{use_author}
\title{Which author to use}
\usage{
use_author(email)
}
\arguments{
\item{email}{An optional email address.
When given and it matches with a single person, the function immediately
returns the information of that person.}
}
\value{
A data.frame with author information.
}
\description{
Reuse existing author information or add a new author.
Allows to update existing author information.
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{menu_first}()},
\code{\link{orcid2person}()},
\code{\link{store_authors}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
