% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_email}
\alias{validate_email}
\title{Check if a vector contains valid email}
\usage{
validate_email(email)
}
\arguments{
\item{email}{A vector with email addresses.}
}
\value{
A logical vector.
}
\description{
It only checks the format of the text, not if the email address exists.
}
\seealso{
Other utils: 
\code{\link{ask_rightsholder_funder}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{install_pak}()},
\code{\link{yesno}()}
}
\concept{utils}
