% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_author.R
\name{validate_orcid}
\alias{validate_orcid}
\title{Validate the structure of an ORCiD id}
\usage{
validate_orcid(orcid)
}
\arguments{
\item{orcid}{A vector of ORCiD}
}
\value{
A logical vector with the same length as the input vector.
}
\description{
Checks whether the ORCiD has the proper format and the checksum.
}
