% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_citation_cff.R
\name{write_citation_cff}
\alias{write_citation_cff}
\title{Write a \code{CITATION.cff} file}
\usage{
write_citation_cff(x = ".", roles)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{roles}{No longer used.}
}
\value{
An invisible \code{checklist} object.
}
\description{
This file format contains the citation information.
It is supported by GitHub, Zenodo and Zotero.
This function is super-seeded by \code{update_citation()}.
}
\seealso{
Other both: 
\code{\link{add_badges}()},
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{custom_dictionary}()},
\code{\link{default_organisation}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{update_citation}()},
\code{\link{write_checklist}()},
\code{\link{write_organisation}()},
\code{\link{write_zenodo_json}()}
}
\concept{both}
