% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_organisation.R
\name{write_organisation}
\alias{write_organisation}
\title{Write organisation settings}
\usage{
write_organisation(org, x = ".")
}
\arguments{
\item{org}{An \code{organisation} object.
Create it with \code{organisation$new()}.}

\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Store the organisation rules into \code{organisation.yml} file.
First run \code{org <- organisation$new()} with the appropriate argument.
Next you can store the configuration with \code{write_organisation(org)}.
}
\seealso{
Other both: 
\code{\link{add_badges}()},
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{custom_dictionary}()},
\code{\link{default_organisation}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{update_citation}()},
\code{\link{write_checklist}()},
\code{\link{write_citation_cff}()},
\code{\link{write_zenodo_json}()}
}
\concept{both}
