% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_zenodo_json.R
\name{write_zenodo_json}
\alias{write_zenodo_json}
\title{Write a \code{.zenodo.json} file}
\usage{
write_zenodo_json(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
An invisible \code{checklist} object.
}
\description{
Zenodo uses the \code{.zenodo.json} file to define the citation information.
See the \href{https://developers.zenodo.org/#add-metadata-to-your-github-repository-release}{Zenodo developers website}
for more information.
}
\seealso{
Other both: 
\code{\link{add_badges}()},
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{custom_dictionary}()},
\code{\link{default_organisation}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{update_citation}()},
\code{\link{write_checklist}()},
\code{\link{write_citation_cff}()},
\code{\link{write_organisation}()}
}
\concept{both}
