% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_badges.R
\name{add_badges}
\alias{add_badges}
\title{add badges to a README}
\usage{
add_badges(x = ".", ...)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{...}{Additional arguments}
}
\description{
\itemize{
\item \code{check_package}: add a package check badge
\item \code{check_project}: add a project check badge
\item \code{doi}: add a DOI badge
\item \code{url}: add a website badge
\item \code{version}: add a version badge
}
}
\examples{
\dontrun{
  add_badges(url = "https://www.inbo.be")
  add_badges(doi = "10.5281/zenodo.8063503")
  add_badges(check_project = "inbo/checklist")
  add_badges(check_package = "inbo/checklist")
  add_badges(version = "v0.1.2")
  add_badges(url = "https://www.inbo.be", doi = "10.5281/zenodo.8063503")
}
}
\seealso{
Other both: 
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{custom_dictionary}()},
\code{\link{default_organisation}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{update_citation}()},
\code{\link{write_checklist}()},
\code{\link{write_citation_cff}()},
\code{\link{write_organisation}()},
\code{\link{write_zenodo_json}()}
}
\concept{both}
