% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_maintainer.R
\name{ask_rightsholder_funder}
\alias{ask_rightsholder_funder}
\title{Ask for rights holder or funder}
\usage{
ask_rightsholder_funder(org, type = c("rightsholder", "funder", "publisher"))
}
\arguments{
\item{org}{Organisation object}

\item{type}{Character, either \code{"rightsholder"}, \code{"funder"} or \code{"publisher"}}
}
\value{
A list with the selected names and the updated organisation object
}
\description{
Ask for rights holder or funder
}
\seealso{
Other utils:
\code{\link[=c_sort]{c_sort()}},
\code{\link[=create_hexsticker]{create_hexsticker()}},
\code{\link[=execshell]{execshell()}},
\code{\link[=install_pak]{install_pak()}},
\code{\link[=yesno]{yesno()}}
}
\concept{utils}
