% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_maintainer.R
\name{ask_rightsholder_funder}
\alias{ask_rightsholder_funder}
\title{Ask for rights holder or funder}
\usage{
ask_rightsholder_funder(org, type = c("rightsholder", "funder"))
}
\arguments{
\item{org}{Organisation object}

\item{type}{Character, either "rightsholder" or "funder"}
}
\value{
A list with the selected names and the updated organisation object
}
\description{
Ask for rights holder or funder
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{install_pak}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
