% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_authors.R
\name{author2df}
\alias{author2df}
\title{Convert person object in a data.frame.}
\usage{
author2df(person)
}
\arguments{
\item{person}{The person object.}
}
\description{
Convert person object in a data.frame.
}
