% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_authors.R
\name{author2df}
\alias{author2df}
\title{Convert person object in a data.frame.}
\usage{
author2df(person)
}
\arguments{
\item{person}{The person object or a list of person objects, \code{NA} or \code{NULL}.
Any \code{"character"} is converted to a person object using \code{as.person()} with a
warning.}
}
\description{
Results in a data.frame with the given name, family name, e-mail, ORCID,
affiliation and role.
Missing elements result in an empty string (\code{""}).
Persons with multiple roles will have the roles as a comma separated string.
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{menu_first}()},
\code{\link{orcid2person}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
