% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_sort.R
\name{c_sort}
\alias{c_sort}
\title{Sort using the C locale}
\usage{
c_sort(x, ...)
}
\arguments{
\item{x}{for \code{sort} an \R object with a class or a numeric,
    complex, character or logical vector.  For \code{sort.int}, a
    numeric, complex, character or logical vector, or a factor.}

\item{...}{arguments to be passed to or from methods or (for the
    default methods and objects without a class) to \code{sort.int}.}
}
\description{
Setting the locale before sorting ensures a stable sorting order
}
\seealso{
Other utils: 
\code{\link{clean_git}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{new_branch}()},
\code{\link{store_authors}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
