% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_codemeta.R
\name{check_codemeta}
\alias{check_codemeta}
\title{Check the package metadata}
\usage{
check_codemeta(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
A \code{Checklist} object.
}
\description{
Check the package metadata
}
\seealso{
Other package: 
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()}
}
\concept{package}
