% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cran.R
\name{check_cran}
\alias{check_cran}
\title{Run the package checks required by CRAN}
\usage{
check_cran(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
A \code{Checklist} object.
}
\description{
Run the package checks required by CRAN
}
