% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cran.R
\name{check_cran}
\alias{check_cran}
\title{Run all the package checks required by CRAN}
\usage{
check_cran(x = ".", quiet = FALSE)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{quiet}{Whether to print check output during checking.}
}
\value{
A \code{checklist} object.
}
\description{
CRAN imposes an impressive list of tests on every package before publication.
This suite of test is available in every R installation.
Hence we use this full suite of tests too.
Notice that \code{check_package()} runs several additional tests.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_environment}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{create_draft_pr}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()},
\code{\link{update_citation}()},
\code{\link{write_citation_cff}()},
\code{\link{write_zenodo_json}()}
}
\concept{package}
