% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cran.R
\name{check_cran}
\alias{check_cran}
\title{Run the package checks required by CRAN}
\usage{
check_cran(path = ".")
}
\arguments{
\item{path}{The path of the package.
Defaults to \code{"."}}
}
\description{
Run the package checks required by CRAN
}
