% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_documentation.R
\name{check_documentation}
\alias{check_documentation}
\title{Check the documentation}
\usage{
check_documentation(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Checks if all required documentation is present and rendered.
A common problem is that one forgets to render the documentation after
updating it.
You'll get an error when you forgot to update the documentation.
But the function will render the documentation while checking it.
So the rendered documentation is available after the check completes.
The function relies on git to detect changes.
Run the checks after committing all changes.
You can amend the commit in case you forgot to render something.
}
\details{
Rules:
\itemize{
\item You must use \href{https://roxygen2.r-lib.org}{roxygen2} to document the
functions.
\item If you use a \code{README.Rmd}, it should be rendered.
You need at least a \code{README.md}.
\item Don't use a \code{NEWS.Rmd} but a \code{NEWS.md}.
\item \code{NEWS.md} must contain an entry for the current package version.
}
}
\section{Required format for \code{NEWS.md}}{
\preformatted{# package_name version

* Description of something that changed.
* Lines should not exceed 80 characters.
  Start a new line with two space to continue an item.
* Add a single blank line before and after each header.

## Second level heading

* You can use second level headings when you want to add more structure.
}
}

\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()}
}
\concept{package}
