% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_environment.R
\name{check_environment}
\alias{check_environment}
\title{Make sure that the required environment variables are set on GitHub}
\usage{
check_environment(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
An invisible \code{checklist} object.
}
\description{
Some actions will fail when these environment variables are not set.
This function does only work on GitHub.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{tidy_desc}()}
}
\concept{package}
