% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{check_files}
\alias{check_files}
\alias{check_allowed_files}
\alias{check_disallowed_files}
\alias{check_required_files}
\title{Check for allowed or disallowed files in a directory}
\usage{
check_allowed_files(
  files = "",
  dir = ".",
  all = FALSE,
  recurse = TRUE,
  type = c("file", "directory", "any"),
  regexp = FALSE
)

check_disallowed_files(files = "", dir = ".", regexp = FALSE)

check_required_files(files = "", dir = ".")
}
\arguments{
\item{files}{Character vector of allowed file names}

\item{dir}{Directory to check}

\item{all}{If \code{TRUE} include hidden files}

\item{recurse}{If \code{TRUE} recurse fully, if a positive number
the number of levels to recurse.}

\item{type}{File type to return, one of "file", "directory", or "any".}

\item{regexp}{If \code{TRUE} use \code{allowed_files} as regexp patterns,
otherwise assume wildcard (glob) patterns}
}
\description{
Check for allowed or disallowed files in a directory
}
\section{Functions}{
\itemize{
\item \code{check_allowed_files}: Check that only the allowed file(s) exist

\item \code{check_disallowed_files}: Check if any disallowed file(s) exist

\item \code{check_required_files}: Check that the required file(s) exist
}}

