% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_description.R
\name{check_license}
\alias{check_license}
\title{Check the license of a package}
\usage{
check_license(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Every package needs a clear license.
Without a license, the end-users have no clue under what conditions they can
use the package.
You must specify the license in the \code{DESCRIPTION} and provide a \code{LICENSE.md}
file.
}
\details{
This functions checks if the \code{DESCRIPTION} mentions one of the standard
licenses.
The \code{LICENSE.md} must match this license.
Use \code{setup_package()} to add the correct \code{LICENSE.md} to the package.

Currently, following licenses are allowed:
\itemize{
\item GPL-3
}

We will consider pull requests adding support for other open source licenses.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_environment}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()},
\code{\link{update_citation}()},
\code{\link{write_citation_cff}()},
\code{\link{write_zenodo_json}()}
}
\concept{package}
