% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lintr.R
\name{check_lintr}
\alias{check_lintr}
\title{Check the packages for linters}
\usage{
check_lintr(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
This functions does \href{https://en.wikipedia.org/wiki/Static_program_analysis}{static code analysis}.
It relies on \code{\link[lintr:lint_package]{lintr::lint_package()}}.
We recommend that you activate all code diagnostics in RStudio to help
meeting the requirements.
You can find this in the menu \emph{Tools} > \emph{Global options} > \emph{Code} >
\emph{Diagnostics}.
}
\details{
Your code must follow the default coding style defined by the lintr package.
\itemize{
\item Use underscore (\verb{_}) to separate long names.
\item No names longer than 30 characters.
\item No lines longer than 80 characters.
\item Use \code{TRUE} and \code{FALSE},
Don't use \code{T} or \code{F}.
\item Add sufficient whitespace characters.
But don't use whitespace when it is not relevant (end of line or file).
\item ...
}
}
\seealso{
Other both: 
\code{\link{check_filename}()},
\code{\link{read_checklist}()},
\code{\link{write_checklist}()}
}
\concept{both}
