% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{A standardised test for R packages}
\usage{
check_package(path = ".")
}
\arguments{
\item{path}{The path of the package.
Defaults to \code{"."}}
}
\description{
A standardised test for R packages
}
