% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{A standardised test for R packages}
\usage{
check_package(x = ".", fail = !interactive())
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{fail}{Should the function return an error in case of a problem?
Defaults to \code{TRUE} on non-interactive session and \code{FALSE} on an interactive
session.}
}
\description{
A standardised test for R packages
}
\seealso{
Other package: 
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()}
}
\concept{package}
