% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkgs.R
\name{check_pkgs}
\alias{check_pkgs}
\alias{installed_pkgs}
\alias{missing_pkgs}
\alias{find_pkgs}
\alias{check_allowed_pkgs}
\alias{check_disallowed_pkgs}
\alias{check_required_pkgs}
\title{Check for allowed or disallowed package dependencies within a project or directory}
\usage{
installed_pkgs()

missing_pkgs(
  dir = here::here(),
  glob = "*.R$|*.r$|*.Rmd$|*.rmd$|*.Rnw$|*.rnw$",
  recurse = TRUE
)

find_pkgs(
  dir = here::here(),
  glob = "*.R$|*.r$|*.Rmd$|*.rmd$|*.Rnw$|*.rnw$",
  full = FALSE,
  recurse = TRUE
)

check_allowed_pkgs(
  pkgs,
  dir = here::here(),
  glob = "*.R$|*.r$|*.Rmd$|*.rmd$|*.Rnw$|*.rnw$",
  full = FALSE,
  recurse = TRUE
)

check_disallowed_pkgs(
  pkgs,
  dir = here::here(),
  glob = "*.R$|*.r$|*.Rmd$|*.rmd$|*.Rnw$|*.rnw$",
  full = FALSE,
  recurse = TRUE
)

check_required_pkgs(
  pkgs,
  dir = here::here(),
  glob = "*.R$|*.r$|*.Rmd$|*.rmd$|*.Rnw$|*.rnw$",
  full = FALSE,
  recurse = TRUE
)
}
\arguments{
\item{dir}{Directory to search.}

\item{glob}{File types to search for, defaults to \code{R}, \code{Rmd}, and \code{Rnw} files.}

\item{recurse}{Should directory be recursively explored (i.e. match files in sub directories)}

\item{full}{Should the full data frame of dependencies be returned or just a vector of package names.}

\item{pkgs}{Character vector of package names}
}
\description{
Check for allowed or disallowed package dependencies within a project or directory
}
\section{Functions}{
\itemize{
\item \code{installed_pkgs}: Returns a vector of installed packages.

\item \code{missing_pkgs}: Returns a vector of packages found by \code{find_pkgs} that are not currently installed.

\item \code{find_pkgs}: Find all of the packages used within a project using renv.

\item \code{check_allowed_pkgs}: Check that only the allowed packages are used

\item \code{check_disallowed_pkgs}: Check if any disallowed packages are used

\item \code{check_required_pkgs}: Check that the required packages are used
}}

