% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rmd.R
\name{check_rmd_renders}
\alias{check_rmd_renders}
\title{Check an Rmd file renders using rmarkdown}
\usage{
check_rmd_renders(file)
}
\arguments{
\item{file}{Path of an Rmd file}
}
\description{
Check an Rmd file renders using rmarkdown
}
