% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_source.R
\name{check_source}
\alias{check_source}
\title{Standardised test for an R source repository}
\usage{
check_source(x = ".", fail = !interactive())
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{fail}{Should the function return an error in case of a problem?
Defaults to \code{TRUE} on non-interactive session and \code{FALSE} on an interactive
session.}
}
\description{
A convenience function that runs all source related tests in sequence.
The details section lists the relevant functions.
When you fixed a problem, you can speed things up by running only the related
check.
We still recommend to run \code{check_source()} before you push to GitHub.
And only push when the functions indicate that there are no problems.
This catches most problems before sending the code to GitHub.
}
\details{
List of checks in order:
\enumerate{
\item \code{check_lintr()}
\item \code{check_filename()}
}
}
\concept{source}
