% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spelling.R
\name{check_spelling}
\alias{check_spelling}
\title{Spell check a package or project}
\usage{
check_spelling(x = ".", quiet = FALSE)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{quiet}{Whether to print check output during checking.}
}
\description{
This function checks by default any markdown (\code{.md}) or Rmarkdown (\code{.Rmd})
file found within the project.
It also checks any R help file (\code{.Rd}) in the \code{man} folder.
Use the \code{set_exceptions()} method of the \code{checklist} object to exclude files
or use a different language.
Have a look at \code{vignette("spelling", package = "checklist")} for more
details.
}
\seealso{
Other both: 
\code{\link{add_badges}()},
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{custom_dictionary}()},
\code{\link{default_organisation}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{update_citation}()},
\code{\link{write_checklist}()},
\code{\link{write_citation_cff}()},
\code{\link{write_organisation}()},
\code{\link{write_zenodo_json}()}
}
\concept{both}
