% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runChecklist.R
\docType{package}
\name{checklist}
\alias{checklist}
\alias{checklist-package}
\title{Sustainable Sourcing Checklist Generator}
\description{
This is a R/Shiny checklist generator app for the sustainable sourcing
project.
}
\examples{
\dontrun{runChecklist()}
}

