% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{checklist}
\alias{checklist}
\title{Obsessive compulsive checks for style and structure}
\usage{
checklist(path = ".")
}
\arguments{
\item{path}{defaults to working directory}
}
\description{
Run a semi-automated checklist on all files at \code{path}.
}
\examples{
# run checklist() on the terrible examples that come with the package:
checklist::checklist(system.file("examples", package = "checklist"))
}
