% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-checklist.R
\name{checklist}
\alias{checklist}
\alias{predict.checklist}
\alias{make_checklist_loss}
\title{Make A Checklist}
\usage{
checklist(x, y, u = 5, simplify = TRUE)

\method{predict}{checklist}(object, x, ...)

make_checklist_loss(u)
}
\arguments{
\item{x}{The independent variables (called features). Typically a matrix.}

\item{y}{The dependent variables. Typically a vector.}

\item{u}{A scale factor of weights assigned to each feature.
Increasing this number will provide greater resolution,
but less interpretability.
Decreasing this number will provide less resolution,
but greater interpretability.}

\item{simplify}{A boolean. Toggles whether to remove coefficients that
have zero weight after being converted to an integer.
Set TRUE by default.}

\item{object}{A checklist object.}

\item{...}{Ignored.}
}
\value{
A checklist object.
}
\description{
Make a checklist.
}
\details{
In checklist, we define a checklist as a linear model with integer
coefficients. When coefficients are integers, it is feasible for a human
to calculate predictions "by hand". When the number of features are few,
the model is easy to interpret and thus easy to implement in practice.
We wrote this tool to make it easy to make a checklist from your data. The tool
will calculate the integer coefficients and eliminate the unimportant features.
Ideally, the features used in a checklist are binary,
but it is up to you to make them binary.
}
