% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_template.R
\name{checklist_template}
\alias{checklist_template}
\title{Create a template with acceptable problems in the package}
\usage{
checklist_template(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the package.
Defaults to \code{.}.}
}
\description{
Create a template with acceptable problems in the package
}
