% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_external_functions.R
\name{cl_load}
\alias{cl_load}
\title{Load a Saved Checlist}
\usage{
cl_load(file)
}
\arguments{
\item{file}{A String containing the file path of the checklist.
Needs to be an .rds file}
}
\value{
A checklist object
}
\description{
This function loads a previously made checklist that was
saved (as an .rds file). It is effectively just a convenience wrapper
of readRDS that will also open your checklist in the Viewer pane
automatically.
}
\examples{
# create a checklist
my_checklist <- cl_create()

# save it
cl_save(my_checklist, "my_checklist.rds")

# remove it from workspace
rm(cl)

# load it back in
cl <- cl_load("my_checklist.rds")

}
