% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_git.R
\name{clean_git}
\alias{clean_git}
\title{Clean the git repository}
\usage{
clean_git(repo = ".", verbose = TRUE)
}
\arguments{
\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link[gert]{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}. When using this
parameter, always explicitly call by name (i.e. \verb{repo = }) because future
versions of gert may have additional parameters.}

\item{verbose}{display some progress info while downloading}
}
\description{
\itemize{
\item update local branches that are behind their counterpart on origin.
\item list local branches that have diverged from their counterpart the origin.
\item list local branches without counterpart on origin that have diverged from
the main branch.
\item remove local branches without counterpart on origin and fully merged into
the main branch.
\item remove local copies of origin branches deleted at the origin.
}
}
\seealso{
Other utils: 
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{new_branch}()},
\code{\link{orcid2person}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
