% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_draft_pr.R
\name{create_draft_pr}
\alias{create_draft_pr}
\title{Create a draft pull request}
\usage{
create_draft_pr(x = ".")
}
\arguments{
\item{x}{The path to the package.}
}
\description{
This function creates a draft pull request for the current version of the
package.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_environment}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()},
\code{\link{update_citation}()},
\code{\link{write_citation_cff}()},
\code{\link{write_zenodo_json}()}
}
\concept{package}
