% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexsticker.R
\name{create_hexsticker}
\alias{create_hexsticker}
\title{Make hexagonal logo for package}
\usage{
create_hexsticker(
  package_name,
  filename = path("man", "figures", "logo.svg"),
  icon,
  x = 0,
  y = 0,
  scale = 1
)
}
\arguments{
\item{package_name}{package name that should be mentioned on the hexagonal
sticker.}

\item{filename}{filename to save the sticker.}

\item{icon}{optional filename to an  \code{.svg} file with an icon.}

\item{x}{number of pixels to move the icon to the right.
Use negative numbers to move the icon to the left.}

\item{y}{number of pixels to move the icon to the bottom.
Use negative numbers to move the icon to the top.}

\item{scale}{Scales the \code{icon}.}
}
\value{
A figure is saved in the working directory or provided path.
}
\description{
This function makes a hexagonal logo in INBO style for the
provided package name.
}
\examples{
\dontrun{
# make tempfile to save logo (or just use (path and) filename)
#' output <- tempfile(pattern = "hexsticker", fileext = ".svg")
create_hexsticker("checklist", filename = output)
}
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{execshell}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{install_pak}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
