% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create an R package according to INBO requirements}
\usage{
create_package(
  package,
  path = ".",
  title,
  description,
  maintainer,
  language = "en-GB",
  license = c("GPL-3", "MIT")
)
}
\arguments{
\item{package}{Name of the new package.}

\item{path}{Where to create the package directory.}

\item{title}{A single sentence with the title of the package.}

\item{description}{A single paragraph describing the package.}

\item{maintainer}{The output of \code{\link[utils:person]{utils::person()}}.
If you use \code{\link[utils:person]{utils::person()}}, then you must provide \code{given}, \code{family},
\code{role}, \code{email} and  \code{comment} with valid \code{ORCID}.
When missing, the functions looks for \code{usethis.description} in the options.
See \code{\link[usethis:use_description]{usethis::use_description()}} for more information.}

\item{language}{Language of the project in \code{xx-YY} format.
\code{xx} is the two letter code for the language.
\code{YY} is the two letter code for the language variant.
E.g. \code{en-GB} for British English, \code{en-US} for American English, \code{nl-BE} for
Belgian Dutch.}

\item{license}{What type of license should be used?
Choice between GPL-3 and MIT.
Default GPL-3.}
}
\description{
Creates a package template in a new folder.
Use this function when you want to start a new package.
Please DO READ \code{vignette("getting_started")} before running this function.
}
\details{
What you get with a checklist setup:
\itemize{
\item minimal folder structure and files required for an R package using INBO
guidelines with GPL-3 or MIT license.
\item an RStudio project file
\item a local git repository
\item an initial \code{NEWS.md} file
\item a template for an \code{README.Rmd}
\item set-up for automated checks and releases of the package using GitHub
Actions.
\item a code of conduct and contributing guidelines.
\item the set-up for a \code{pkgdown} website using the INBO corporate identity.
}
}
\examples{
# maintainer in `utils::person()` format
maintainer <- person(
  given = "Thierry",
  family = "Onkelinx",
  role = c("aut", "cre"),
  email = "thierry.onkelinx@inbo.be",
  comment = c(ORCID = "0000-0001-8804-4216")
)

# creating the package
path <- tempfile()
dir.create(path)
create_package(
  path = path, package = "packagename", title = "package title",
  description = "A short description.", maintainer = maintainer,
  language = "en-GB", license = "GPL-3"
)
}
\seealso{
Other setup: 
\code{\link{create_project}()},
\code{\link{prepare_ghpages}()},
\code{\link{setup_package}()},
\code{\link{setup_project}()},
\code{\link{setup_source}()}
}
\concept{setup}
