% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_dictionary.R
\name{custom_dictionary}
\alias{custom_dictionary}
\title{Add words to custom dictionaries}
\usage{
custom_dictionary(issues)
}
\arguments{
\item{issues}{The output of \code{check_spelling()}.}
}
\description{
Add words to custom dictionaries
}
\seealso{
Other both: 
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{print.checklist_spelling}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{write_checklist}()},
\code{\link{write_organisation}()}
}
\concept{both}
