% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{orcid2person}
\alias{orcid2person}
\title{Defunct functions}
\usage{
orcid2person()
}
\description{
Defunct functions
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
