% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execshell}
\alias{execshell}
\title{Pass command lines to a shell}
\usage{
execshell(commandstring, intern = FALSE, path = ".", ...)
}
\arguments{
\item{commandstring}{The system command to be invoked, as a string.
Multiple commands can be combined in this single string, e.g. with a
multiline string.}

\item{intern}{a logical (not \code{NA}) which indicates whether to
    capture the output of the command as an \R character vector.}

\item{path}{The path from where the command string needs to be executed}

\item{...}{Other arguments passed to \code{\link[base:system]{base::system()}} or
\code{base::shell()}.}
}
\description{
Cross-platform function to pass a command to the shell, using either
\code{\link[base:system]{base::system()}} or (Windows-only) \code{base::shell()}, depending on the
operating system.
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{menu_first}()},
\code{\link{orcid2person}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
