% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pak.R
\name{install_pak}
\alias{install_pak}
\title{Install extra packages defined in \code{checklist.yml}}
\usage{
install_pak(x = ".", ...)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{...}{Additional arguments passed to \code{pak::pkg_install()}}
}
\description{
Install extra packages defined in \code{checklist.yml}
}
