% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pak.R
\name{install_pak}
\alias{install_pak}
\title{Install extra packages defined in \code{checklist.yml}}
\usage{
install_pak(x = ".", ...)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{...}{Additional arguments passed to \code{pak::pkg_install()}}
}
\description{
Install extra packages defined in \code{checklist.yml}
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
