% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_workdir_clean}
\alias{is_workdir_clean}
\title{Check if the current working directory of a repo is clean}
\usage{
is_workdir_clean(repo)
}
\arguments{
\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link[gert]{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}. When using this
parameter, always explicitly call by name (i.e. \verb{repo = }) because future
versions of gert may have additional parameters.}
}
\value{
\code{TRUE} when there are no staged, unstaged or untracked files.
Otherwise \code{FALSE}
}
\description{
A clean working directory has no staged, unstaged or untracked files.
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{clean_git}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{is_repository}()},
\code{\link{menu_first}()},
\code{\link{new_branch}()},
\code{\link{orcid2person}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
