% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_branch.R
\name{new_branch}
\alias{new_branch}
\title{Create a new branch after cleaning the repo}
\usage{
new_branch(branch, verbose = TRUE, checkout = TRUE, repo = ".")
}
\arguments{
\item{branch}{name of branch to check out}

\item{verbose}{display some progress info while downloading}

\item{checkout}{move HEAD to the newly created branch}

\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link[gert]{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}. When using this
parameter, always explicitly call by name (i.e. \verb{repo = }) because future
versions of gert may have additional parameters.}
}
\description{
This functions first runs \code{clean_git()}.
Then it creates the new branch from the (updated) main branch.
}
\seealso{
Other git: 
\code{\link{clean_git}()},
\code{\link{create_draft_pr}()},
\code{\link{get_branches_tags}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{set_tag}()}
}
\concept{git}
