% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_org_list.R
\name{new_org_list}
\alias{new_org_list}
\title{Interactively create a new organisation list}
\usage{
new_org_list(git)
}
\arguments{
\item{git}{An optional string with the absolute path to a git
organisation.
E.g. \code{"https://github.com/inbo"}}
}
\description{
An interactive alternative for \code{org_list$new()}.
Reuses available organisations where possible.
}
\seealso{
\code{\link{org_list}}, \code{\link{org_item}}

Other setup: 
\code{\link{create_package}()},
\code{\link{create_project}()},
\code{\link{prepare_ghpages}()},
\code{\link{set_license}()},
\code{\link{setup_package}()},
\code{\link{setup_project}()},
\code{\link{setup_source}()}
}
\concept{setup}
