% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orcid2person}
\alias{orcid2person}
\title{Convert an ORCID to a \code{person} object.}
\usage{
orcid2person(orcid, email, role = c("aut", "cre"))
}
\arguments{
\item{orcid}{The ORCID of the person.}

\item{email}{An optional email of the person.
Require when the ORCID record does not contain a public email.}

\item{role}{The role of the person.
See \code{utils::person} for all possible values.}
}
\description{
This function requires that your \code{ORCID_TOKEN} is set as an environment
variable.
First run \code{rorcid::orcid_auth()}.
A browser window should open where you can log into \code{ORCID}.
Run \code{rorcid::orcid_auth()}, which should return something like
\code{"Bearer dc0a6b6b-b4d4-4276-bc89-78c1e9ede56e"}.
Copy this (not the \code{Bearer} part) and append it to your \code{.Renviron} as
follows: \verb{ORCID_TOKEN=dc0a6b6b-b4d4-4276-bc89-78c1e9ede56e}.
Don't forget to append your UUID instead of the example given here.
}
\seealso{
Other utils: 
\code{\link{c_sort}()},
\code{\link{clean_git}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{new_branch}()},
\code{\link{store_authors}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
