% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_list_class.R
\name{org_list}
\alias{org_list}
\title{The \code{org_list}  R6 class}
\description{
A class containing a list of organisations
}
\seealso{
Other class: 
\code{\link{checklist}},
\code{\link{citation_meta}},
\code{\link{org_item}},
\code{\link{organisation}},
\code{\link{spelling}}
}
\concept{class}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-org_list-new}{\code{org_list$new()}}
\item \href{#method-org_list-print}{\code{org_list$print()}}
\item \href{#method-org_list-read}{\code{org_list$read()}}
\item \href{#method-org_list-write}{\code{org_list$write()}}
\item \href{#method-org_list-clone}{\code{org_list$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-new"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{org_list} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more \code{org_item} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-print"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{org_list} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-read"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-read}{}}}
\subsection{Method \code{read()}}{
Read the \code{org_list} object from an \code{organisation.yml} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$read(x = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The path to the directory where the \code{organisation.yml} file
is stored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-write"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-write}{}}}
\subsection{Method \code{write()}}{
Write the \code{org_list} object to an \code{organisation.yml} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$write(x = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The path to the directory where the \code{organisation.yml} file
should be written.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-org_list-clone"></a>}}
\if{latex}{\out{\hypertarget{method-org_list-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{org_list$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
