% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spelling.R
\name{print.checklist_spelling}
\alias{print.checklist_spelling}
\title{Display a \code{checklist_spelling} summary}
\usage{
\method{print}{checklist_spelling}(x, ...)
}
\arguments{
\item{x}{The \code{checklist_spelling} object}

\item{...}{currently ignored}
}
\description{
Display a \code{checklist_spelling} summary
}
\seealso{
Other both: 
\code{\link{add_badges}()},
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{custom_dictionary}()},
\code{\link{default_organisation}()},
\code{\link{read_checklist}()},
\code{\link{read_organisation}()},
\code{\link{write_checklist}()},
\code{\link{write_organisation}()}
}
\concept{both}
