% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit_on_failure.R
\name{quit_on_failure}
\alias{quit_on_failure}
\title{Quit on Failure}
\usage{
quit_on_failure(expr)
}
\arguments{
\item{expr}{Expression returning a logical vector}
}
\value{
expr
}
\description{
Evaluate the given expression and if any of the returned values are \code{FALSE} then
exit with status 1.
}
\details{
This is needed when using \verb{check_*} functions within a continuous integration pipeline
to signal that a step failed.
}
