% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_checklist.R
\name{read_checklist}
\alias{read_checklist}
\title{Read the check list file from a package}
\usage{
read_checklist(path = ".")
}
\arguments{
\item{path}{The path of the package.
Defaults to \code{"."}}
}
\value{
a named list with check list information
}
\description{
Read the check list file from a package
}
