% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_checklist.R
\name{read_checklist}
\alias{read_checklist}
\title{Read the check list file from a package}
\usage{
read_checklist(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
A \code{Checklist} object.
}
\description{
Read the check list file from a package
}
\seealso{
Other both: 
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{write_checklist}()}
}
\concept{both}
