% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_checklist.R
\name{read_checklist}
\alias{read_checklist}
\title{Read the check list file from a package}
\usage{
read_checklist(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
A \code{checklist} object.
}
\description{
The checklist package stores configuration information in the \code{checklist.yml}
file in the root of a project.
This function reads this configuration.
It is mainly used by the other functions inside the package.
}
\seealso{
Other both: 
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{check_spelling}()},
\code{\link{print.checklist_spelling}()},
\code{\link{write_checklist}()}
}
\concept{both}
