% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tag.R
\name{set_tag}
\alias{set_tag}
\title{Set a new tag}
\usage{
set_tag(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Set a new tag
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_package}()},
\code{\link{tidy_desc}()}
}
\concept{package}
