% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tag.R
\name{set_tag}
\alias{set_tag}
\title{Set a new tag}
\usage{
set_tag(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Set a new tag
}
