% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_package.R
\name{setup_package}
\alias{setup_package}
\title{Add or update the checklist infrastructure to an existing package}
\usage{
setup_package(path = ".")
}
\arguments{
\item{path}{The path to the package.
Defaults to \code{"."}.}
}
\description{
Use this function when you have an existing package and you want to use the
checklist functionality.
Please keep in mind that the checklist is an opinionated list of checks.
It might require some breaking changes.
DO READ \code{vignette("getting_started")} before you run this function.
}
\details{
What you get with a checklist setup:
\itemize{
\item minimal folder structure and files required for an R package using INBO
guidelines with GPL-3 license.
\item an RStudio project file
\item a local git repository
\item an initial \code{NEWS.md} file
\item a template for an \code{README.Rmd}
\item set-up for automate checks and releases of the package using GitHub
Actions.
\item a code of conduct and contributing guidelines.
\item the set-up for a \code{pkgdown} website using the INBO corporate identity.
}
}
\seealso{
Other setup: 
\code{\link{create_package}()},
\code{\link{setup_source}()}
}
\concept{setup}
