% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_source.R
\name{setup_source}
\alias{setup_source}
\title{Add or update the checklist infrastructure to a repository with source files.}
\usage{
setup_source(path = ".")
}
\arguments{
\item{path}{The path to the project.
Defaults to \code{"."}.}
}
\description{
This adds the required GitHub workflows to run \code{check_source()} automatically
whenever you push commits to GitHub.
It also add a \href{https://creativecommons.org/licenses/by-sa/4.0/}{CC-BY-SA 4.0}
license file, a \href{https://inbo.github.io/checklist/CODE_OF_CONDUCT.html}{CODE_OF_CONDUCT.md} and the checklist
configuration file (\code{checklist.yml}).
}
\seealso{
Other setup: 
\code{\link{create_package}()},
\code{\link{setup_package}()}
}
\concept{setup}
