% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_source.R
\name{setup_source}
\alias{setup_source}
\title{Add or update the checklist infrastructure to a repository with source files.}
\usage{
setup_source(path = ".", language = "en-GB")
}
\arguments{
\item{path}{The path to the project.
Defaults to \code{"."}.}

\item{language}{Language of the project in \code{xx-YY} format.
\code{xx} is the two letter code for the language.
\code{YY} is the two letter code for the language variant.
E.g. \code{en-GB} for British English, \code{en-US} for American English, \code{nl-BE} for
Belgian Dutch.}
}
\description{
This adds the required GitHub workflows to run \code{check_source()} automatically
whenever you push commits to GitHub.
It also adds a
\href{https://creativecommons.org/licenses/by/4.0/}{CC-BY 4.0} license file,
a
\href{https://inbo.github.io/checklist/CODE_OF_CONDUCT.html}{\code{CODE_OF_CONDUCT.md}}
and the checklist configuration file (\code{checklist.yml}).
}
\seealso{
Other setup: 
\code{\link{create_package}()},
\code{\link{create_project}()},
\code{\link{prepare_ghpages}()},
\code{\link{set_license}()},
\code{\link{setup_package}()},
\code{\link{setup_project}()}
}
\concept{setup}
