% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_authors.R
\name{store_authors}
\alias{store_authors}
\title{Store author details for later usage}
\usage{
store_authors(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Store author details for later usage
}
\seealso{
Other utils: 
\code{\link{c_sort}()},
\code{\link{clean_git}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{new_branch}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
