% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_description.R
\name{tidy_desc}
\alias{tidy_desc}
\title{Make your DESCRIPTION tidy}
\usage{
tidy_desc(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
A tidy \code{DESCRIPTION} uses a strict formatting and order of key-value pairs.
This function reads the current \code{DESCRIPTION} and overwrites it with a tidy
version.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_environment}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{update_citation}()},
\code{\link{write_citation_cff}()},
\code{\link{write_zenodo_json}()}
}
\concept{package}
