% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_description.R
\name{tidy_desc}
\alias{tidy_desc}
\title{Make your description tidy}
\usage{
tidy_desc(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
Make your description tidy
}
