% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_description.R
\name{tidy_desc}
\alias{tidy_desc}
\title{Make your DESCRIPTION tidy}
\usage{
tidy_desc(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\description{
A tidy \code{DESCRIPTION} use a strict formatting and order of key-value pairs.
This function reads the current \code{DESCRIPTION} and overwrites is with a tidy
version.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_license}()},
\code{\link{check_package}()}
}
\concept{package}
