% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_citation.R
\name{update_citation}
\alias{update_citation}
\title{Create or update the citation files}
\usage{
update_citation(x = ".", quiet = FALSE)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
An invisible \code{checklist} object.
}
\description{
The function extracts citation meta data from the project.
Then it checks the required meta data.
Upon success, it writes several files.
}
\details{
\itemize{
\item \code{.zenodo.json} contains the citation information in the format that
\href{https://zenodo.org}{Zenodo} requires.
\item \code{CITATION.cff} provides the citation information in the format that
\href{https://github.com}{GitHub} requires.
\item \code{inst/CITATION} provides the citation information in the format that
R packages require.
It is only relevant for packages.
}
}
\note{
Source of the citation meta data:
\itemize{
\item package: Mainly \code{DESCRIPTION}.
keywords and (Zenodo) community extracted from \code{README.md}.
\item project: \code{README.md}
}

Should you want to add more information to the \code{inst/CITATION} file,
add it to that file outside \verb{# begin checklist entry} and
\verb{# end checklist entry}.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_environment}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()},
\code{\link{write_citation_cff}()},
\code{\link{write_zenodo_json}()}
}
\concept{package}
