% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_citation.R
\name{update_citation}
\alias{update_citation}
\title{Create or update the CITATION file}
\usage{
update_citation(x = ".", roles)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{roles}{Roles to select the persons for the \code{DESCRIPTION}.
Defaults to \code{c("aut", "cre")}.}
}
\value{
An invisible \code{checklist} object.
}
\description{
The function creates the file \code{inst/CITATION} is case is doesn't exist.
Otherwise it will update the citation information indicated between
\verb{# begin checklist entry} and \verb{# end checklist entry}.
The remained of the file is left untouched.
}
\details{
The content of \code{DESCRIPTION} determines the citation information.

When you don't want the update, then remove both \verb{# begin checklist entry}
and \verb{# end checklist entry}.
Note that this will result in a warning, which you can allow via
\code{write_checklist()}.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_environment}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()},
\code{\link{write_citation_cff}()},
\code{\link{write_zenodo_json}()}
}
\concept{package}
