% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_author.R
\name{use_author}
\alias{use_author}
\title{Which author to use}
\usage{
use_author(email, lang)
}
\arguments{
\item{email}{An optional email address.
When given and it matches with a single person, the function immediately
returns the information of that person.}

\item{lang}{The language to use for the affiliation.
Defaults to the first language in the \code{name} vector of the
\code{org_list} object.
When the affiliation is not available in that language,
it will use the first available language.}
}
\value{
A data.frame with author information.
}
\description{
Reuse existing author information or add a new author.
Allows to update existing author information.
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{install_pak}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{validate_email}()},
\code{\link{validate_orcid}()},
\code{\link{yesno}()}
}
\concept{utils}
