% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_email}
\alias{validate_email}
\title{Check if a vector contains valid email}
\usage{
validate_email(email)
}
\arguments{
\item{email}{A vector with email addresses.}
}
\value{
A logical vector.
}
\description{
It only checks the format of the text, not if the email works.
}
\seealso{
Other utils: 
\code{\link{is_workdir_clean}()},
\code{\link{orcid2person}()},
\code{\link{yesno}()}
}
\concept{utils}
