% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_email}
\alias{validate_email}
\title{Check if a vector contains valid email}
\usage{
validate_email(email)
}
\arguments{
\item{email}{A vector with email addresses.}
}
\value{
A logical vector.
}
\description{
It only checks the format of the text, not if the email address exists.
}
\seealso{
Other utils: 
\code{\link{ask_yes_no}()},
\code{\link{c_sort}()},
\code{\link{clean_git}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{is_repository}()},
\code{\link{is_workdir_clean}()},
\code{\link{menu_first}()},
\code{\link{new_branch}()},
\code{\link{orcid2person}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{yesno}()}
}
\concept{utils}
