% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_author.R
\name{validate_orcid}
\alias{validate_orcid}
\title{Validate the structure of an ORCID id}
\usage{
validate_orcid(orcid)
}
\arguments{
\item{orcid}{A vector of ORCID}
}
\value{
A logical vector with the same length as the input vector.
}
\description{
Checks whether the ORCID has the proper format and the checksum.
}
\seealso{
Other utils: 
\code{\link{ask_rightsholder_funder}()},
\code{\link{ask_yes_no}()},
\code{\link{author2df}()},
\code{\link{bookdown_zenodo}()},
\code{\link{c_sort}()},
\code{\link{create_hexsticker}()},
\code{\link{execshell}()},
\code{\link{get_default_org_list}()},
\code{\link{inbo_org_list}()},
\code{\link{install_pak}()},
\code{\link{menu_first}()},
\code{\link{store_authors}()},
\code{\link{use_author}()},
\code{\link{validate_email}()},
\code{\link{yesno}()}
}
\concept{utils}
