% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_checklist.R
\name{write_checklist}
\alias{write_checklist}
\title{Write a check list with allowed issues in the source code}
\usage{
write_checklist(x = ".", package = TRUE)
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{package}{Logical indication if \code{checklist.yml} refers to an R package.
Defaults to \code{TRUE}.}
}
\description{
First run \code{x <- checklist::check_package()} or
\code{x <- checklist::check_source()}.
These commands run the checks and store the CheckList object in the variable
\code{x}.
Next you can store the configuration with \code{checklist::write_checklist(x)}.
This will first list any existing allowed warnings or notes.
For every one of them, choose whether you want to keep it or not.
Next, the function presents every new warning or note which you may allow or
not.
If you choose to allow a warning or note, you must provide a motivation.
Please provide a sensible motivation.
Keep in mind that \code{checklist.yml} stores these motivations in plain text,
so they are visible for other users.
We use the \code{yesno()} function to make sure you carefully read the questions.
}
\seealso{
Other both: 
\code{\link{check_filename}()},
\code{\link{check_lintr}()},
\code{\link{read_checklist}()}
}
\concept{both}
