% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_checklist.R
\name{write_checklist}
\alias{write_checklist}
\title{Write a check list with allowed issues in the package}
\usage{
write_checklist(x = ".")
}
\arguments{
\item{x}{Either a \code{Checklist} object or a path to the package.
Defaults to \code{.}.}
}
\description{
Write a check list with allowed issues in the package
}
