% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_citation_cff.R
\name{write_citation_cff}
\alias{write_citation_cff}
\title{Write a \code{CITATION.cff} file}
\usage{
write_citation_cff(x = ".", roles)
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}

\item{roles}{Roles to select the persons for the \code{DESCRIPTION}.
Defaults to \code{c("aut", "cre")}.}
}
\value{
An invisible \code{checklist} object.
}
\description{
This file format contains the citation information.
It is supported by GitHub, Zenodo and Zotero.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_environment}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()},
\code{\link{update_citation}()},
\code{\link{write_zenodo_json}()}
}
\concept{package}
