% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_zenodo_json.R
\name{write_zenodo_json}
\alias{write_zenodo_json}
\title{Write a \code{.zenodo.json} file}
\usage{
write_zenodo_json(x = ".")
}
\arguments{
\item{x}{Either a \code{checklist} object or a path to the source code.
Defaults to \code{.}.}
}
\value{
An invisible \code{checklist} object.
}
\description{
Zenodo uses the \code{.zenodo.json} file to define the citation information.
See the \href{https://developers.zenodo.org/#add-metadata-to-your-github-repository-release}{Zenodo developers website}
for more information.
}
\seealso{
Other package: 
\code{\link{check_codemeta}()},
\code{\link{check_cran}()},
\code{\link{check_description}()},
\code{\link{check_documentation}()},
\code{\link{check_environment}()},
\code{\link{check_license}()},
\code{\link{check_package}()},
\code{\link{set_tag}()},
\code{\link{tidy_desc}()},
\code{\link{update_citation}()},
\code{\link{write_citation_cff}()}
}
\concept{package}
